static methods instead of DI
